pyproject.toml
[project]
name = "agentx"
version = "2.0.0"
description = "Arena-based tree-search ML agent for MLE-Bench competitions"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "uvicorn>=0.27.0",
    "anthropic>=0.52.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.1.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "optuna>=3.4.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
